refactor: Menu, MenuOption 등 엔티티에 BaseEntity 상속 적용 및 공통 필드 이전

+++ src/main/java/com/sparta/tentenbackend/domain/menu/dto/MenuDto.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class MenuDto {
   private UUID id;
   private String name;
-  private BigInteger price;
+  private Long price;
   private String description;
 
   public static MenuDto fromEntity(Menu menu) {
